찜목록 디비에 넣음

+++ app/src/main/java/com/example/mu338/stampinseoul/ThemeData.java
@@ -9,8 +9,8 @@ public class ThemeData implements Serializable {
     private String addr;
     private String tel;
     private String overView;
-    private long mapX;
-    private long mapY;
+    private double mapX;
+    private double mapY;
     private boolean hart=false;
 
     public boolean isHart() {
@@ -39,19 +39,19 @@ public class ThemeData implements Serializable {
         this.tel = tel;
     }
 
-    public long getMapX() {
+    public double getMapX() {
         return mapX;
     }
 
-    public void setMapX(long mapX) {
+    public void setMapX(double mapX) {
         this.mapX = mapX;
     }
 
-    public long getMapY() {
+    public double getMapY() {
         return mapY;
     }
 
-    public void setMapY(long mapY) {
+    public void setMapY(double mapY) {
         this.mapY = mapY;
     }
